SB#16 Merge update SML partition state
